refactor :: core:data AuthRepository import package 변경 및 스레드 추가
